requirements.txt
# Core ML and Vision
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
huggingface-hub>=0.16.0
tokenizers>=0.13.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0

# Image Processing
Pillow>=9.0.0
opencv-python-headless>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0
dill>=0.3.7
fsspec[http]<=2025.3.0,>=2023.1.0
scikit-learn>=1.3.0

# Web Interface
gradio>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Scalability and Production
ray[default]==2.49.1
redis==5.0.1
psutil==5.9.6
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# FastAPI and Async Processing
fastapi>=0.104.0
celery[redis]>=5.3.0
python-multipart>=0.0.6

# Large-scale Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
parquet-python>=1.3.1

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0